% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scGSEA.R
\name{add.scGESA.to.obj}
\alias{add.scGESA.to.obj}
\title{Add scGESA results to single cell object}
\usage{
add.scGESA.to.obj(sc.obj, gs.obj)
}
\value{
a `SeqGeneSet` object with data added to slot sc.ES.perm and sc.normFlag reset as TRUE
}
\description{
This functions adds scGESA results to single cell object
}
\examples{
\dontrun{
}
}
\author{
Wei Sheng (wsheng at stu.njmu.edu.cn)
}
