% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scGSEA.R
\name{getGSgenelist}
\alias{getGSgenelist}
\title{Get gene list from a `SeqGeneSet` object}
\usage{
getGSgenelist(gs.obj, GeneSetName)
}
\arguments{
\item{gs.obj}{a `SeqGeneSet` object}

\item{GeneSetName}{Name(s) of geneSet(s)}
}
\value{
a vector of gene names
}
\description{
This function extracts a list of genes in geneSet(s) specificed by GeneSetName.
}
\examples{
\dontrun{
}
}
\author{
Wei Sheng (wsheng at stu.njmu.edu.cn)
}
