% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scGSEA.R
\name{scCalES}
\alias{scCalES}
\title{Calculate ES score}
\usage{
scCalES(sc.obj, gs.obj, weighted.type = 0)
}
\arguments{
\item{sc.obj}{either a `Seurat` or a `SingleCellExperiment` object}

\item{gs.obj}{a `SeqGeneSet` object}

\item{weighted.type}{the weighted type in GSEA}
}
\value{
a `SeqGeneSet` object with data added to slot sc.ES
}
\description{
This function is used to calcuate ES score for each cell and each gene set
}
\examples{
\dontrun{
}
}
\author{
Wei Sheng (wsheng at stu.njmu.edu.cn)
}
