% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scGSEA.R
\name{scCalSignif}
\alias{scCalSignif}
\title{Calculate ES significance}
\usage{
scCalSignif(gs.obj)
}
\value{
a `SeqGeneSet` object with data added to slots sc.pval, sc.FDR, and sc.FWER
}
\description{
This function calculates ES significance, including p-value, FDR, and FWER
}
\examples{
\dontrun{
}
}
\author{
Wei Sheng (wsheng at stu.njmu.edu.cn)
}
