% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scGSEA.R
\name{scESnorm}
\alias{scESnorm}
\title{ES score normalizaton}
\usage{
scESnorm(
  sc.obj,
  gs.obj,
  weighted.type = 0,
  perm.time = ncol(sc.obj@scale.data),
  seed = 1234
)
}
\arguments{
\item{sc.obj}{either a `Seurat` or a `SingleCellExperiment` object}

\item{gs.obj}{a `SeqGeneSet` object}

\item{weighted.type}{the weighted type in GSEA}

\item{perm.time}{the number of permutations, with default value the number of cells}

\item{seed}{the seed for random sampling}
}
\value{
a `SeqGeneSet` object with data added to slot sc.ES.perm and sc.normFlag reset as TRUE
}
\description{
This function calculates ES scores for a permutated data sets, 
which is then used for normalization of the observed ES scores
}
\examples{
\dontrun{
}
}
\author{
Wei Sheng (wsheng at stu.njmu.edu.cn)
}
