% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scGSEA.R
\name{scGSEA.scale.data}
\alias{scGSEA.scale.data}
\title{Scale gene expression for scGSEA}
\usage{
scGSEA.scale.data(obj, gene_expr_frac = 0.1, kNN = 21)
}
\arguments{
\item{obj}{either a `Seurat` or a `SingleCellExperiment` object}

\item{gene_expr_frac}{fliltering out genes if the non-zero fraction less than this number}

\item{kNN}{the k nearest neighbors for data smoothing/denoising.}
}
\value{
the obj with added data in slot `assay` of `Seurat` and `metadata` of `SingleCellExperiment`
}
\description{
This function normalizes single cell gene expression levels acorss cells for each gene.
}
\examples{
obj <- scGSEA.scale.data(obj, kNN=51)
\dontrun{
}
}
\author{
Wei Sheng (wsheng at stu.njmu.edu.cn)
}
