% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scGSEA.R
\name{scLoadGS}
\alias{scLoadGS}
\title{Load and store gene sets for scGSEA}
\usage{
scLoadGS(
  geneset.file,
  sc.obj,
  geneID.type = c("gene.symbol", "ensembl"),
  use.HVG = FALSE,
  genesetsize.min = 5,
  genesetsize.max = 1000
)
}
\arguments{
\item{geneset.file}{the file path and name of GeneSet in .gmt format}

\item{sc.obj}{either a `Seurat` or a `SingleCellExperiment` object}

\item{geneID.type}{the gene ID type, currently support gene symbol and Ensembl}

\item{use.HVG}{logical, using highly variable genes only or not}

\item{genesetsize.min}{minimum number of genes in a gene set, gene set with smaller than this number of genes will be excluded}

\item{genesetsize.max}{maximum number of genes in a gene set, gene set with greater than this number of genes will be excluded}
}
\value{
a `SeqGeneSet` object
}
\description{
This function loads gene sets and creats a `SeqGeneSet` object
}
\examples{
\dontrun{
}
}
\author{
Wei Sheng (wsheng at stu.njmu.edu.cn)
}
